microbial_deep_learning/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
astroid==2.2.5
atomicwrites==1.1.5
attrs==19.1.0
awscli==1.16.172
Babel==2.7.0
backcall==0.1.0
bidict==0.18.0
biopython==1.71
bleach==3.1.0
botocore==1.12.162
chardet==3.0.4
click==6.7
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
Deprecated==1.2.5
dill==0.2.9
docutils==0.14
entrypoints==0.3
enum34==1.1.6
fancycompleter==0.8
flake8==3.7.7
Flask==1.0.2
idna==2.6
imagesize==1.1.0
importlib-metadata==0.17
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
itsdangerous==0.24
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
more-itertools==4.1.0
multiprocess==0.70.7
names==0.3.0
nbformat==4.4.0
numpy==1.16.4
packaging==19.0
pandas==0.24.2
parso==0.4.0
pathlib2==2.3.3
pathos==0.2.3
pdbpp==0.10.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==3.10.0
pluggy==0.12.0
pox==0.2.5
ppft==1.6.4.9
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytorch-ignite==0.2.0
pytz==2019.1
PyYAML==3.13
pyzmq==18.0.1
requests==2.22.0
retrying==1.3.3
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
simplegeneric==0.8.1
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.33.11
torch==1.1.0
torchfile==0.1.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
wmctrl==0.3
wrapt==1.11.1
zipp==0.5.1
fireworks-ml==0.3.1
ipdb